Now multiresolution mel-spectrogram supports multichannel. Started working on multichannel convolution.

+++ scripts/audio_functions/signal_generation.py
@@ -16,7 +16,7 @@ def create_log_sweep(length: int, f0: float, f1: float, sr: float, silence: int,
     t = np.arange(0, int(length * sr)) / sr
     sweep = chirp(t, f0=f0, f1=f1, t1=length, method='logarithmic')
 
-    sweep = np.concatenate([sweep, np.zeros(silence * sr)], axis=0)
+    sweep = np.concatenate([sweep, np.zeros(int(silence * sr))], axis=0)
 
     if n_channels > 1:
         sweep = np.stack([sweep] * n_channels)
